Reduce innate length

+++ apps/cc3o3/elements/ElementFwd.h
@@ -6,9 +6,9 @@ namespace RF::cc::element {
 ///////////////////////////////////////////////////////////////////////////////
 
 // Element and innate identifier bytes are intended to be valid ASCII strings
-using ElementIdentifier = uint64_t;
+using ElementIdentifier = uint64_t; // 'B' 'u' 'f' 'S' '_' 'R' '\0' '\0'
 static constexpr ElementIdentifier kInvalidElementIdentifier = 0;
-using InnateIdentifier = uint64_t;
+using InnateIdentifier = uint32_t; // 'r' 'e' 'd' '\0'
 static constexpr InnateIdentifier kInvalidInnateIdentifier = 0;
 
 using ElementLevel = uint8_t;
